+++ framework/camera.hpp
@@ -12,7 +12,7 @@ class Camera{
 public:
 	Camera();
 	Camera(glm::vec3 const& eye_, glm::vec3 const& direction_, glm::vec3 const& up_, float fov_degrees);
-	Ray generate_ray_at(glm::ivec2 const& screen_coord, int rayDepth)const;
+	Ray generate_ray_at(glm::vec2 const& screen_coord, int rayDepth)const;
 
 	void translate(glm::vec3 const& translationVector);
 	void rotate(float degrees, glm::vec3 const& axis);
